refactor(map): Move previews logic in separate package
